Add the JNI method to create a CA signed certificate (#404)
This is labeled as insecure because that's exactly what it is.  The CA key is right there in the function call, so the user will need a really good reason if they are going to use it in production, but it's useful for testing and for environments that are secured by other means.

This method is very similar to the self signed generate method, so I refactored it into another common method that will vary based on if caKey and caCertificate are both non null or not.  If they are both non null, they get passed to signRequest instead of the subject key and null.

The result will be a certificate that is signed by an Android key store key and issued by the provided CA.